fix and check other features
