+++ Makefile
@@ -2,6 +2,11 @@ test:
 	cargo fmt -- --check
 	cargo test
 	cargo build --features deny_warnings
+	cargo check --no-default-features --features deny_warnings
+	cargo check --no-default-features --features deny_warnings --features pcap
+	cargo check --no-default-features --features deny_warnings --features node
+	cargo check --no-default-features --features deny_warnings --features rpc_client
+	cargo check --no-default-features --features deny_warnings --features rpc_server
 	cargo run --example cli --features deny_warnings -- target/debug/feeless
 
 fix:
